# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Музыкальная библиотека",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/song": {
            "get": {
                "tags": [
                    "Песни"
                ],
                "summary": "Найти песню",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Исполнитель. Игнорирует регистр букв, для совпадения хватит части имени/названия группы",
                        "name": "group",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Страница. Нумерация начинается с 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "type": "integer",
                        "description": "Размер страницы. Не может быть больше 100 или меньше 0",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Даты должны быть в формате YYYY-MM-DD",
                        "name": "releasedAfter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Даты должны быть в формате YYYY-MM-DD",
                        "name": "releasedBefore",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Название песни. Игнорирует регистр букв, для совпадения хватит части названия",
                        "name": "song",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Текст песни. Игнорирует регистр букв, для совпадения хватит части тектса",
                        "name": "text",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Song"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Песни"
                ],
                "summary": "Добавить песню по группе и названию",
                "parameters": [
                    {
                        "description": "Название и исполнитель",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/PostRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Песни"
                ],
                "summary": "Удалить песню по Id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID песни",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Песни"
                ],
                "summary": "Исправить данные песни",
                "parameters": [
                    {
                        "type": "string",
                        "name": "group",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "link",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "releaseDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "song",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "text",
                        "in": "query"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/song/{id}": {
            "get": {
                "tags": [
                    "Песни"
                ],
                "summary": "Найти песню по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID песни",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/song/{id}/{part}": {
            "get": {
                "tags": [
                    "Песни"
                ],
                "summary": "Найти фрагмент слов песни по ID и номеру фрагмента",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID песни",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Номер фрагмента",
                        "name": "part",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "PostRequest": {
            "type": "object",
            "required": [
                "group",
                "song"
            ],
            "properties": {
                "group": {
                    "type": "string",
                    "example": "Nothing More"
                },
                "song": {
                    "type": "string",
                    "example": "Angel song"
                }
            }
        },
        "Song": {
            "type": "object",
            "properties": {
                "group": {
                    "type": "string",
                    "example": "Nothing More"
                },
                "id": {
                    "type": "integer"
                },
                "link": {
                    "type": "string",
                    "example": "https://www.youtube.com/watch?v=WLJ9b6HIMHw"
                },
                "releaseDate": {
                    "type": "string",
                    "example": "2024-05-17"
                },
                "song": {
                    "type": "string",
                    "example": "Angel Song"
                },
                "text": {
                    "type": "string",
                    "example": "(Hey, hey, hey)\n(Hey, hey, hey)\nHey!\n\nYou can't deny you got that feeling in your bones\nReady to go\nAre you ready to go?"
                }
            }
        }
    }
}